rename tests2.rs -> tests.rs and bench_2.rs -> bench.rs
