+++ Cargo.toml
@@ -23,7 +23,7 @@ serde_json = "1.0"
 criterion = "0.3"
 
 [[bench]]
-name = "bench_2"
+name = "bench"
 harness = false
 
 [features]
