+++ src/lib.rs
@@ -52,7 +52,7 @@
 //! ```
 
 #[cfg(test)]
-mod tests2;
+mod tests;
 
 #[cfg(feature = "serde-1")]
 mod serde_impl;
